Move garbage collect code into storage package
